+++ src/config.rs
@@ -15,34 +15,49 @@
 
 //! Configuration for git-z.
 
+pub mod updater;
+
 mod v0_1;
-
-use std::{fs, io, path::PathBuf, process::Command};
-
-use thiserror::Error;
+mod v0_2_dev_0;
 
 // NOTE: When you switch to a new version:
 //
 // - write a new version module,
 // - switch the version here,
 // - update VERSION below,
-// - write a new `impl From<old::Config> for Config` implementations,
-// - write a new load_from_* method to load from the previous version,
-// - write an updater in `git-z update`,
-// - update this message with instructions for next versions.
-pub use v0_1::Config;
+// - update the version in `git-z.toml[.sample]`,
+// - update the `impl From<old::Config> for Config` implementations,
+// - write a new `impl From<previous::Config> for Config` implementation,
+// - handle the previous config in `Config::load`,
+// - write an updater in `ConfigUpdater`,
+// - update the previous updaters as well,
+// - update `git z update`.
+pub use v0_2_dev_0::{Config, Scopes, Templates, Ticket};
 
-/// An error that can occur when loading the config.
+use std::{fs, io, path::PathBuf, process::Command};
+
+use indexmap::IndexMap;
+use serde::{Deserialize, Serialize};
+use thiserror::Error;
+
+/// An error that can occur when loading the configuration.
 #[derive(Debug, Error)]
 pub enum LoadError {
-    #[error("Failed to get the config file path")]
+    #[error("Failed to get the configuration file path")]
     ConfigFileError(#[from] ConfigFileError),
     #[error("Failed to read {CONFIG_FILE_NAME}")]
     ReadError(#[from] io::Error),
-    #[error("Failed to parse {CONFIG_FILE_NAME}")]
+    #[error("Invalid configuration in {CONFIG_FILE_NAME}")]
+    InvalidConfig(#[from] FromTomlError),
+}
+
+/// An error that can occur when parsing the TOML.
+#[derive(Debug, Error)]
+pub enum FromTomlError {
+    #[error("Configuration version {0} is not supported")]
+    UnsupportedVersion(String),
+    #[error("Failed to parse into a valid configuration")]
     ParseError(#[from] toml::de::Error),
-    #[error("The config file is out of date")]
-    OutOfDate,
 }
 
 /// An error that can occur when building the config file path.
@@ -63,22 +78,42 @@ pub enum RepoRootError {
     EncodingError(#[from] std::string::FromUtf8Error),
 }
 
-const CONFIG_FILE_NAME: &str = "git-z.toml";
-const VERSION: &str = "0.1";
+/// A minimal configuration to get the version.
+///
+/// The configuration format for git-z can evolve with time. It is versioned for
+/// this purpose, so that git-z is able to select the proper parser. This struct
+/// allows to parse any configuration as long as it contains a version field.
+#[derive(Debug, Serialize, Deserialize)]
+struct MinimalConfig {
+    /// The version of the configuration.
+    version: String,
+}
+
+/// The name of the configuration file.
+pub const CONFIG_FILE_NAME: &str = "git-z.toml";
+/// The current version of the configuration file.
+pub const VERSION: &str = "0.2-dev.0";
 
 const DEFAULT_TEMPLATE: &str = include_str!("../templates/COMMIT_EDITMSG");
 
 impl Default for Config {
     fn default() -> Self {
         Self {
-            version: String::from("0.1"),
-            types: vec![
-                String::from("feat  introduces a new feature"),
-                String::from("fix   patches a bug"),
-            ],
-            scopes: vec![],
-            template: String::from(DEFAULT_TEMPLATE),
-            ticket_prefixes: vec![String::from("")],
+            version: String::from(VERSION),
+            types: IndexMap::from([
+                (
+                    String::from("feat"),
+                    String::from("introduces a new feature"),
+                ),
+                (String::from("fix"), String::from("patches a bug")),
+            ]),
+            scopes: None,
+            ticket: Ticket {
+                prefixes: vec![String::from("")],
+            },
+            templates: Templates {
+                commit: String::from(DEFAULT_TEMPLATE),
+            },
         }
     }
 }
@@ -86,19 +121,29 @@ impl Default for Config {
 impl Config {
     /// Loads the configuration the repo or fallbacks to the default.
     pub fn load() -> Result<Self, LoadError> {
-        let config = match fs::read_to_string(config_file()?) {
-            Ok(config) => toml::from_str(&config)?,
+        match fs::read_to_string(config_file()?) {
+            Ok(config) => Ok(Self::from_toml(&config)?),
 
             Err(error) => match error.kind() {
-                io::ErrorKind::NotFound => Self::default(),
-                _ => return Err(LoadError::ReadError(error)),
+                io::ErrorKind::NotFound => Ok(Self::default()),
+                _ => Err(LoadError::ReadError(error)),
             },
-        };
+        }
+    }
 
-        if config.version == VERSION {
-            Ok(config)
-        } else {
-            Err(LoadError::OutOfDate)
+    /// Builds the configuration from its TOML representation.
+    pub fn from_toml(toml: &str) -> Result<Self, FromTomlError> {
+        let minimal_config: MinimalConfig = toml::from_str(toml)?;
+
+        match minimal_config.version.as_str() {
+            VERSION => Ok(toml::from_str(toml)?),
+            "0.1" => {
+                let config: v0_1::Config = toml::from_str(toml)?;
+                Ok(config.into())
+            }
+            version => {
+                Err(FromTomlError::UnsupportedVersion(version.to_owned()))
+            }
         }
     }
 }
@@ -119,4 +164,35 @@ fn repo_root() -> Result<PathBuf, RepoRootError> {
         let git_error = String::from_utf8(git_rev_parse.stderr)?;
         Err(RepoRootError::GitError(git_error.trim().to_owned()))
     }
+}
+
+impl From<v0_1::Config> for Config {
+    fn from(old: v0_1::Config) -> Self {
+        Self {
+            version: old.version,
+            types: split_types_and_docs(old.types),
+            scopes: Some(Scopes::List { list: old.scopes }),
+            ticket: Ticket {
+                prefixes: old.ticket_prefixes,
+            },
+            templates: Templates {
+                commit: old.template,
+            },
+        }
+    }
+}
+
+fn split_types_and_docs(types: Vec<String>) -> IndexMap<String, String> {
+    types
+        .iter()
+        .map(AsRef::as_ref)
+        .map(split_type_and_doc)
+        .collect()
+}
+
+fn split_type_and_doc(type_and_doc: &str) -> (String, String) {
+    let mut split = type_and_doc.splitn(2, ' ');
+    let ty = split.next().unwrap_or_default().to_owned();
+    let doc = split.next().unwrap_or_default().trim().to_owned();
+    (ty, doc)
 }